clean up Economy and change event register order
